add restores view
