+++ assets/src/routes/backupRoutes.tsx
@@ -3,13 +3,19 @@ import { Navigate, Route } from 'react-router-dom'
 import Root from '../components/backups/Backups'
 import Clusters from '../components/backups/clusters/Clusters'
 import ObjectStores from '../components/backups/objectstores/ObjectStores'
+
 import Cluster from '../components/backups/cluster/Cluster'
+import ClusterBackups from '../components/backups/cluster/backups/Backups'
+import ClusterRestores from '../components/backups/cluster/restores/Restores'
 
 import {
   BACKUPS_ABS_PATH,
   BACKUPS_DEFAULT_REL_PATH,
   CLUSTERS_REL_PATH,
   CLUSTER_ABS_PATH,
+  CLUSTER_BACKUPS_DEFAULT_REL_PATH,
+  CLUSTER_BACKUPS_REL_PATH,
+  CLUSTER_RESTORES_REL_PATH,
   OBJECT_STORES_REL_PATH,
 } from './backupRoutesConsts'
 
@@ -39,5 +45,23 @@ export const backupsRoutes = [
   <Route
     path={CLUSTER_ABS_PATH}
     element={<Cluster />}
-  />,
+  >
+    <Route
+      index
+      element={
+        <Navigate
+          replace
+          to={CLUSTER_BACKUPS_DEFAULT_REL_PATH}
+        />
+      }
+    />
+    <Route
+      path={CLUSTER_BACKUPS_REL_PATH}
+      element={<ClusterBackups />}
+    />
+    <Route
+      path={CLUSTER_RESTORES_REL_PATH}
+      element={<ClusterRestores />}
+    />
+  </Route>,
 ]
